Add files via upload
Visit the *131# on your phone and write a program that displays an exact output as shown o your phone. Ensure additionally that in your program If a user chooses option 0, there is a different window that's loaded. among the options in the new window is a 0, which takes a user back to the previous page when entered. Paste the GitHub link here

Signed-off-by: IanLusule <134999927+IanLusule@users.noreply.github.com>